Refactor PDF delegate classes into their own view
